能动

+++ src/main/java/com/zz/pojo/STUDENT.java
@@ -4,15 +4,16 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class STUDENT {
-    Integer SNO;
+    BigDecimal SNO;
     String sname;
     String ssex;
     String sbirthday;
     String speciality;
     String SCLASS;
-    Integer TC;
+    BigDecimal TC;
 }
